dashboard/requirements.txt
airtable==0.4.5
dash==2.10.0
dash-bootstrap-components==1.4.1
pandas==2.0.3
plotly==5.15.0
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.3.0
numpy==1.24.3
firebase-admin==6.2.0 